+++ app/repositories/AuthorsRepository.php
@@ -2,11 +2,9 @@
 
 namespace App\Repositories;
 
-use App\Core\Database;
-
-class AuthorsRepository
+class AuthorsRepository extends Repository
 {
-    public static function all(Database $db)
+    public static function all()
     {
         $query = 'select a.id, a.name, a.slug, a.email, a.profile_picture, a.created_at, COUNT(p.author_id) as posts_count
             from authors as a
@@ -15,13 +13,13 @@ class AuthorsRepository
             GROUP BY a.id
             ORDER BY a.created_at';
 
-        return $db->query($query)->all();
+        return self::db()->query($query)->all();
     }
 
-    public static function get(Database $db, string $slug)
+    public static function get(string $slug)
     {
         $query = 'select * from authors where slug = :slug';
 
-        return $db->query($query, compact('slug'))->findOrFail();
+        return self::db()->query($query, compact('slug'))->findOrFail();
     }
 }