+++ views/home.view.php
@@ -4,7 +4,7 @@
     <div class="flex justify-between container mx-auto">
         <div class="w-full lg:w-8/12">
             <div class="flex items-center justify-between">
-                <h1 class="text-xl font-bold text-gray-700 md:text-2xl">Featured articles <?= $queries['q'] ?? '' ?></h1>
+                <h1 class="text-xl font-bold text-gray-700 md:text-2xl">Featured articles</h1>
                 <?php component("post-filter") ?>
             </div>
             <?php foreach ($posts as $post) : ?>